Configurable light strip code; interfaces with RIO
Just now realized this should be on the GIT repo. Should communicate
with RoboRIO via digital inputs. We'll be able to change color depending
on alliance in competition as well. Each robot can support up to two
1-meter long strips, powered by the 5V1.5A plugs on the VRM. The Arduino
will be USB powered, and the light patterns are configurable, via drive
station. May cause problems while disabled, but shouldn't be a too bad
(in theory it will just default to the BlueGreenPurple pattern).